option_set factory
